requirements.txt
stable-worldmodel[train,env]
duckietown-gym-daffy
h5py
einops
matplotlib
scikit-learn
xvfbwrapper
